split header files and source files
