+++ modules/mcc/include/opencv2/mcc.hpp
@@ -32,6 +32,8 @@
 
 #include "mcc/checker_detector.hpp"
 #include "mcc/checker_model.hpp"
+#include "mcc/ccm.hpp"
+
 
 /** @defgroup mcc Macbeth Chart module
