+++ modules/mcc/include/opencv2/mcc/colorspace.hpp
@@ -32,9 +32,9 @@
 #include <vector>
 #include <string>
 #include <iostream>
-#include "opencv2/mcc/io.hpp"
-#include "opencv2/mcc/operations.hpp"
-#include "opencv2/mcc/utils.hpp"
+#include "io.hpp"
+#include "operations.hpp"
+//#include "utils.hpp"
 
 namespace cv
 {
@@ -43,16 +43,16 @@ namespace ccm
 
 /* *\ brief Basic class for ColorSpace.
 */
-class ColorSpace
+class CV_EXPORTS ColorSpace
 {
 public:
-    IO io;
-    std::string type;
-    bool linear;
-    Operations to;
-    Operations from;
-    ColorSpace* l = 0;
-    ColorSpace* nl = 0;
+    CV_PROP_RW IO io;
+    CV_PROP_RW std::string type;
+    CV_PROP_RW bool linear;
+    CV_PROP_RW Operations to;
+    CV_PROP_RW Operations from;
+    CV_PROP_RW ColorSpace* l = 0;
+    CV_PROP_RW ColorSpace* nl = 0;
 
     ColorSpace() {};
 
@@ -64,40 +64,31 @@ public:
         nl = 0;
     };
 
-    virtual bool relate(const ColorSpace& other) const
-    {
-        return (type == other.type) && (io == other.io);
-    };
+    CV_WRAP virtual bool relate(const ColorSpace& other) const;
 
-    virtual Operations relation(const ColorSpace& /*other*/) const
-    {
-        return IDENTITY_OPS;
-    };
+    CV_WRAP virtual Operations relation(const ColorSpace& /*other*/) const;
 
-    bool operator<(const ColorSpace& other)const
-    {
-        return (io < other.io || (io == other.io && type < other.type) || (io == other.io && type == other.type && linear < other.linear));
-    }
+    bool operator<(const ColorSpace& other)const;
 };
 
 /* *\ brief Base of RGB color space;
    *        the argument values are from AdobeRGB;
    *        Data from https://en.wikipedia.org/wiki/Adobe_RGB_color_space
 */
-class RGBBase_ : public ColorSpace
+class CV_EXPORTS RGBBase_ : public ColorSpace
 {
 public:
-    // primaries
-    double xr;
-    double yr;
-    double xg;
-    double yg;
-    double xb;
-    double yb;
-    MatFunc toL;
-    MatFunc fromL;
-    cv::Mat M_to;
-    cv::Mat M_from;
+    primaries
+    CV_PROP_RW double xr;
+    CV_PROP_RW double yr;
+    CV_PROP_RW double xg;
+    CV_PROP_RW double yg;
+    CV_PROP_RW double xb;
+    CV_PROP_RW double yb;
+    CV_PROP_RW MatFunc toL;
+    CV_PROP_RW MatFunc fromL;
+    CV_PROP_RW cv::Mat M_to;
+    CV_PROP_RW cv::Mat M_from;
 
     using ColorSpace::ColorSpace;
 
@@ -109,41 +100,16 @@ public:
        *\ param other type of ColorSpace.
        *\ return Operations.
     */
-    Operations relation(const ColorSpace& other) const CV_OVERRIDE
-    {
-        if (linear == other.linear)
-        {
-            return IDENTITY_OPS;
-        }
-        if (linear)
-        {
-            return Operations({ Operation(fromL) });
-        }
-        return Operations({ Operation(toL) });
-    };
+    CV_WRAP Operations relation(const ColorSpace& other) const CV_OVERRIDE;
 
     /* *\ brief Initial operations.
     */
-    void init()
-    {
-        setParameter();
-        calLinear();
-        calM();
-        calOperations();
-    }
+    CV_WRAP void init();
 
     /* *\ brief Produce color space instance with linear and non-linear versions.
        *\ param rgbl type of RGBBase_.
     */
-    void bind(RGBBase_& rgbl)
-    {
-        init();
-        rgbl.init();
-        l = &rgbl;
-        rgbl.l = &rgbl;
-        nl = this;
-        rgbl.nl = this;
-    }
+    CV_WRAP void bind(RGBBase_& rgbl);
 
 private:
     virtual void setParameter() {};
@@ -151,362 +117,190 @@ private:
     /* *\ brief Calculation of M_RGBL2XYZ_base.
        *        see ColorSpace.pdf for details.
     */
-    virtual void calM()
-    {
-        cv::Mat XYZr, XYZg, XYZb, XYZ_rgbl, Srgb;
-        XYZr = cv::Mat(xyY2XYZ({ xr, yr }), true);
-        XYZg = cv::Mat(xyY2XYZ({ xg, yg }), true);
-        XYZb = cv::Mat(xyY2XYZ({ xb, yb }), true);
-        merge(std::vector<cv::Mat>{ XYZr, XYZg, XYZb }, XYZ_rgbl);
-        XYZ_rgbl = XYZ_rgbl.reshape(1, XYZ_rgbl.rows);
-        cv::Mat XYZw = cv::Mat(illuminants.find(io)->second, true);
-        solve(XYZ_rgbl, XYZw, Srgb);
-        merge(std::vector<cv::Mat>{ Srgb.at<double>(0)* XYZr,
-            Srgb.at<double>(1)* XYZg,
-            Srgb.at<double>(2)* XYZb }, M_to);
-        M_to = M_to.reshape(1, M_to.rows);
-        M_from = M_to.inv();
-    };
+    virtual void calM();
 
     /* *\ brief operations to or from XYZ.
     */
-    virtual void calOperations()
-    {
-        // rgb -> rgbl
-        toL = [this](cv::Mat rgb)->cv::Mat {return toLFunc(rgb); };
-
-        // rgbl -> rgb
-        fromL = [this](cv::Mat rgbl)->cv::Mat {return fromLFunc(rgbl); };
-
-        if (linear)
-        {
-            to = Operations({ Operation(M_to.t()) });
-            from = Operations({ Operation(M_from.t()) });
-        }
-        else
-        {
-            to = Operations({ Operation(toL), Operation(M_to.t()) });
-            from = Operations({ Operation(M_from.t()), Operation(fromL) });
-        }
-    }
+    virtual void calOperations();
 
     virtual void calLinear() {}
 
-    virtual cv::Mat toLFunc(cv::Mat& /*rgb*/)
-    {
-        return cv::Mat();
-    };
+    virtual cv::Mat toLFunc(cv::Mat& /*rgb*/);
 
-    virtual cv::Mat fromLFunc(cv::Mat& /*rgbl*/)
-    {
-        return cv::Mat();
-    };
+    virtual cv::Mat fromLFunc(cv::Mat& /*rgbl*/);
 
 };
 
 /* *\ brief Base of Adobe RGB color space;
 */
-class AdobeRGBBase_ : public RGBBase_
+class CV_EXPORTS AdobeRGBBase_ : public RGBBase_
 {
 public:
     using RGBBase_::RGBBase_;
-    double gamma;
+     double gamma;
 
 private:
-    cv::Mat toLFunc(cv::Mat& rgb) CV_OVERRIDE
-    {
-        return gammaCorrection(rgb, gamma);
-    }
-
-    cv::Mat fromLFunc(cv::Mat& rgbl) CV_OVERRIDE
-    {
-        return gammaCorrection(rgbl, 1. / gamma);
-    }
+    cv::Mat toLFunc(cv::Mat& rgb) CV_OVERRIDE;
+    cv::Mat fromLFunc(cv::Mat& rgbl) CV_OVERRIDE;
 };
 
 /* *\ brief Base of sRGB color space;
 */
-class sRGBBase_ : public RGBBase_
+class CV_EXPORTS sRGBBase_ : public RGBBase_
 {
 public:
     using RGBBase_::RGBBase_;
-    double a;
-    double gamma;
-    double alpha;
-    double beta;
-    double phi;
-    double K0;
+    CV_PROP_RW double a;
+    CV_PROP_RW double gamma;
+    CV_PROP_RW double alpha;
+    CV_PROP_RW double beta;
+    CV_PROP_RW double phi;
+    CV_PROP_RW double K0;
 
 private:
     /* *\ brief linearization parameters
        *        see ColorSpace.pdf for details;
     */
-    virtual void calLinear() CV_OVERRIDE
-    {
-        alpha = a + 1;
-        K0 = a / (gamma - 1);
-        phi = (pow(alpha, gamma) * pow(gamma - 1, gamma - 1)) / (pow(a, gamma - 1) * pow(gamma, gamma));
-        beta = K0 / phi;
-    }
+    virtual void calLinear() CV_OVERRIDE;
 
     /* *\ brief Used by toLFunc.
     */
-    double toLFuncEW(double& x)
-    {
-        if (x > K0)
-        {
-            return pow(((x + alpha - 1) / alpha), gamma);
-        }
-        else if (x >= -K0)
-        {
-            return x / phi;
-        }
-        else
-        {
-            return -(pow(((-x + alpha - 1) / alpha), gamma));
-        }
-    }
+    double toLFuncEW(double& x);
 
     /* *\ brief Linearization.
        *        see ColorSpace.pdf for details.
        *\ param rgb the input array, type of cv::Mat.
        *\ return the output array, type of cv::Mat.
     */
-    cv::Mat toLFunc(cv::Mat& rgb) CV_OVERRIDE
-    {
-        return elementWise(rgb, [this](double a_)->double {return toLFuncEW(a_); });
-    }
+    cv::Mat toLFunc(cv::Mat& rgb) CV_OVERRIDE;
 
     /* *\ brief Used by fromLFunc.
     */
-    double fromLFuncEW(double& x)
-    {
-        if (x > beta)
-        {
-            return alpha * pow(x, 1 / gamma) - (alpha - 1);
-        }
-        else if (x >= -beta)
-        {
-            return x * phi;
-        }
-        else
-        {
-            return -(alpha * pow(-x, 1 / gamma) - (alpha - 1));
-        }
-    }
+    double fromLFuncEW(double& x);
 
     /* *\ brief Delinearization.
        *        see ColorSpace.pdf for details.
        *\ param rgbl the input array, type of cv::Mat.
        *\ return the output array, type of cv::Mat.
     */
-    cv::Mat fromLFunc(cv::Mat& rgbl) CV_OVERRIDE
-    {
-        return elementWise(rgbl, [this](double a_)->double {return fromLFuncEW(a_); });
-    }
+    cv::Mat fromLFunc(cv::Mat& rgbl) CV_OVERRIDE;
 };
 
 /* *\ brief sRGB color space.
    *        data from https://en.wikipedia.org/wiki/SRGB.
 */
-class sRGB_ :public sRGBBase_
+class CV_EXPORTS sRGB_ :public sRGBBase_
 {
 public:
     sRGB_(bool linear_) :sRGBBase_(D65_2, "sRGB", linear_) {};
 
 private:
-    void setParameter() CV_OVERRIDE
-    {
-        xr = 0.64;
-        yr = 0.33;
-        xg = 0.3;
-        yg = 0.6;
-        xb = 0.15;
-        yb = 0.06;
-        a = 0.055;
-        gamma = 2.4;
-    }
+    void setParameter() CV_OVERRIDE;
 };
 
 /* *\ brief Adobe RGB color space.
 */
-class AdobeRGB_ : public AdobeRGBBase_
+class CV_EXPORTS AdobeRGB_ : public AdobeRGBBase_
 {
 public:
     AdobeRGB_(bool linear_ = false) :AdobeRGBBase_(D65_2, "AdobeRGB", linear_) {};
 
 private:
-    void setParameter() CV_OVERRIDE
-    {
-        xr = 0.64;
-        yr = 0.33;
-        xg = 0.21;
-        yg = 0.71;
-        xb = 0.15;
-        yb = 0.06;
-        gamma = 2.2;
-    }
+    void setParameter() CV_OVERRIDE;
 };
 
 /* *\ brief Wide-gamut RGB color space.
    *        data from https://en.wikipedia.org/wiki/Wide-gamut_RGB_color_space.
 */
-class WideGamutRGB_ : public AdobeRGBBase_
+class CV_EXPORTS WideGamutRGB_ : public AdobeRGBBase_
 {
 public:
     WideGamutRGB_(bool linear_ = false) :AdobeRGBBase_(D50_2, "WideGamutRGB", linear_) {};
 
 private:
-    void setParameter() CV_OVERRIDE
-    {
-        xr = 0.7347;
-        yr = 0.2653;
-        xg = 0.1152;
-        yg = 0.8264;
-        xb = 0.1566;
-        yb = 0.0177;
-        gamma = 2.2;
-    }
+    void setParameter() CV_OVERRIDE;
 };
 
 /* *\ brief ProPhoto RGB color space.
    *        data from https://en.wikipedia.org/wiki/ProPhoto_RGB_color_space.
 */
-class ProPhotoRGB_ : public AdobeRGBBase_
+class CV_EXPORTS ProPhotoRGB_ : public AdobeRGBBase_
 {
 public:
     ProPhotoRGB_(bool linear_ = false) :AdobeRGBBase_(D50_2, "ProPhotoRGB", linear_) {};
 
 private:
-    void setParameter() CV_OVERRIDE
-    {
-        xr = 0.734699;
-        yr = 0.265301;
-        xg = 0.159597;
-        yg = 0.840403;
-        xb = 0.036598;
-        yb = 0.000105;
-        gamma = 1.8;
-    }
+    void setParameter() CV_OVERRIDE;
 };
 
 /* *\ brief DCI-P3 RGB color space.
    *        data from https://en.wikipedia.org/wiki/DCI-P3.
 */
-class DCI_P3_RGB_ : public AdobeRGBBase_
+class CV_EXPORTS DCI_P3_RGB_ : public AdobeRGBBase_
 {
 public:
     DCI_P3_RGB_(bool linear_ = false) :AdobeRGBBase_(D65_2, "DCI_P3_RGB", linear_) {};
 
 private:
-    void setParameter() CV_OVERRIDE
-    {
-        xr = 0.68;
-        yr = 0.32;
-        xg = 0.265;
-        yg = 0.69;
-        xb = 0.15;
-        yb = 0.06;
-        gamma = 2.2;
-    }
+    void setParameter() CV_OVERRIDE;
 };
 
 /* *\ brief Apple RGB color space.
    *        data from http://www.brucelindbloom.com/index.html?WorkingSpaceInfo.html.
 */
-class AppleRGB_ : public AdobeRGBBase_
+class CV_EXPORTS AppleRGB_ : public AdobeRGBBase_
 {
 public:
     AppleRGB_(bool linear_ = false) :AdobeRGBBase_(D65_2, "AppleRGB", linear_) {};
 
 private:
-    void setParameter() CV_OVERRIDE
-    {
-        xr = 0.625;
-        yr = 0.34;
-        xg = 0.28;
-        yg = 0.595;
-        xb = 0.155;
-        yb = 0.07;
-        gamma = 1.8;
-    }
+    void setParameter() CV_OVERRIDE;
 };
 
 /* *\ brief REC_709 RGB color space.
    *        data from https://en.wikipedia.org/wiki/Rec._709.
 */
-class REC_709_RGB_ : public sRGBBase_
+class CV_EXPORTS REC_709_RGB_ : public sRGBBase_
 {
 public:
     REC_709_RGB_(bool linear_) :sRGBBase_(D65_2, "REC_709_RGB", linear_) {};
 
 private:
-    void setParameter() CV_OVERRIDE
-    {
-        xr = 0.64;
-        yr = 0.33;
-        xg = 0.3;
-        yg = 0.6;
-        xb = 0.15;
-        yb = 0.06;
-        a = 0.099;
-        gamma = 1 / 0.45;
-    }
+    void setParameter() CV_OVERRIDE;
 };
 
 /* *\ brief REC_2020 RGB color space.
    *        data from https://en.wikipedia.org/wiki/Rec._2020.
 */
-class REC_2020_RGB_ : public sRGBBase_
+class CV_EXPORTS REC_2020_RGB_ : public sRGBBase_
 {
 public:
     REC_2020_RGB_(bool linear_) :sRGBBase_(D65_2, "REC_2020_RGB", linear_) {};
 
 private:
-    void setParameter() CV_OVERRIDE
-    {
-        xr = 0.708;
-        yr = 0.292;
-        xg = 0.17;
-        yg = 0.797;
-        xb = 0.131;
-        yb = 0.046;
-        a = 0.09929682680944;
-        gamma = 1 / 0.45;
-    }
+    void setParameter() CV_OVERRIDE;
 };
 
 
-sRGB_ sRGB(false), sRGBL(true);
-AdobeRGB_ AdobeRGB(false), AdobeRGBL(true);
-WideGamutRGB_ WideGamutRGB(false), WideGamutRGBL(true);
-ProPhotoRGB_ ProPhotoRGB(false), ProPhotoRGBL(true);
-DCI_P3_RGB_ DCI_P3_RGB(false), DCI_P3_RGBL(true);
-AppleRGB_ AppleRGB(false), AppleRGBL(true);
-REC_709_RGB_ REC_709_RGB(false), REC_709_RGBL(true);
-REC_2020_RGB_ REC_2020_RGB(false), REC_2020_RGBL(true);
+extern sRGB_ sRGB, sRGBL;
+extern AdobeRGB_ AdobeRGB, AdobeRGBL;
+extern WideGamutRGB_ WideGamutRGB, WideGamutRGBL;
+extern ProPhotoRGB_ ProPhotoRGB, ProPhotoRGBL;
+extern DCI_P3_RGB_ DCI_P3_RGB, DCI_P3_RGBL;
+extern AppleRGB_ AppleRGB, AppleRGBL;
+extern REC_709_RGB_ REC_709_RGB, REC_709_RGBL;
+extern REC_2020_RGB_ REC_2020_RGB, REC_2020_RGBL;
+
 
 /* *\ brief Bind RGB with RGBL.
 */
-class ColorSpaceInitial
+class CV_EXPORTS ColorSpaceInitial
 {
 public:
-    ColorSpaceInitial()
-    {
-        sRGB.bind(sRGBL);
-        AdobeRGB.bind(AdobeRGBL);
-        WideGamutRGB.bind(WideGamutRGBL);
-        ProPhotoRGB.bind(ProPhotoRGBL);
-        DCI_P3_RGB.bind(DCI_P3_RGBL);
-        AppleRGB.bind(AppleRGBL);
-        REC_709_RGB.bind(REC_709_RGBL);
-        REC_2020_RGB.bind(REC_2020_RGBL);
-
-    }
+    CV_WRAP ColorSpaceInitial();
 };
 
-ColorSpaceInitial color_space_initial;
-
+extern ColorSpaceInitial color_space_initial;
 
 /* *\ brief Enum of the possible types of CAMs.
 */
@@ -529,14 +323,11 @@ const static std::map <CAM, std::vector< cv::Mat >> MAs = {
 /* *\ brief XYZ color space.
    *        Chromatic adaption matrices.
 */
-class XYZ :public ColorSpace
+class CV_EXPORTS XYZ :public ColorSpace
 {
 public:
     XYZ(IO io_) : ColorSpace(io_, "XYZ", true) {};
-    Operations cam(IO dio, CAM method = BRADFORD)
-    {
-        return (io == dio) ? Operations() : Operations({ Operation(cam_(io, dio, method).t()) });
-    }
+    CV_WRAP Operations cam(IO dio, CAM method = BRADFORD);
 
 private:
     /* *\ brief Get cam.
@@ -545,27 +336,7 @@ private:
        *\ param method type of CAM.
        *\ return the output array, type of cv::Mat.
     */
-    cv::Mat cam_(IO sio, IO dio, CAM method = BRADFORD) const
-    {
-        if (sio == dio)
-        {
-            return cv::Mat::eye(cv::Size(3, 3), CV_64FC1);
-        }
-        if (cams.count(std::make_tuple(dio, sio, method)) == 1)
-        {
-            return cams[std::make_tuple(dio, sio, method)];
-        }
-
-        // Function from http ://www.brucelindbloom.com/index.html?ColorCheckerRGB.html.
-        cv::Mat XYZws = cv::Mat(illuminants.find(dio)->second);
-        cv::Mat XYZWd = cv::Mat(illuminants.find(sio)->second);
-        cv::Mat MA = MAs.at(method)[0];
-        cv::Mat MA_inv = MAs.at(method)[1];
-        cv::Mat M = MA_inv * cv::Mat::diag((MA * XYZws) / (MA * XYZWd)) * MA;
-        cams[std::make_tuple(dio, sio, method)] = M;
-        cams[std::make_tuple(sio, dio, method)] = M.inv();
-        return M;
-    }
+    cv::Mat cam_(IO sio, IO dio, CAM method = BRADFORD) const;
 };
 
 /* *\ brief Define XYZ_D65_2 and XYZ_D50_2.
@@ -576,14 +347,10 @@ const XYZ XYZ_D50_2(D50_2);
 
 /* *\ brief Lab color space.
 */
-class Lab :public ColorSpace
+class CV_EXPORTS Lab :public ColorSpace
 {
 public:
-    Lab(IO io_) : ColorSpace(io_, "XYZ", true)
-    {
-        to = { Operation([this](cv::Mat src)->cv::Mat {return tosrc(src); }) };
-        from = { Operation([this](cv::Mat src)->cv::Mat {return fromsrc(src); }) };
-    }
+    Lab(IO io_);
 
 private:
     static constexpr double delta = (6. / 29.);
@@ -591,38 +358,21 @@ private:
     static constexpr double t0 = delta * delta * delta;
     static constexpr double c = 4. / 29.;
 
-    cv::Vec3d fromxyz(cv::Vec3d& xyz)
-    {
-        double x = xyz[0] / illuminants.find(io)->second[0], y = xyz[1] / illuminants.find(io)->second[1], z = xyz[2] / illuminants.find(io)->second[2];
-        auto f = [this](double t)->double { return t > t0 ? std::cbrtl(t) : (m * t + c); };
-        double fx = f(x), fy = f(y), fz = f(z);
-        return { 116. * fy - 16. ,500 * (fx - fy),200 * (fy - fz) };
-    }
+    cv::Vec3d fromxyz(cv::Vec3d& xyz);
 
     /* *\ brief Calculate From.
        *\ param src the input array, type of cv::Mat.
        *\ return the output array, type of cv::Mat
     */
-    cv::Mat fromsrc(cv::Mat& src)
-    {
-        return channelWise(src, [this](cv::Vec3d a)->cv::Vec3d {return fromxyz(a); });
-    }
+    cv::Mat fromsrc(cv::Mat& src);
 
-    cv::Vec3d tolab(cv::Vec3d& lab)
-    {
-        auto f_inv = [this](double t)->double {return t > delta ? pow(t, 3.0) : (t - c) / m; };
-        double L = (lab[0] + 16.) / 116., a = lab[1] / 500., b = lab[2] / 200.;
-        return { illuminants.find(io)->second[0] * f_inv(L + a),illuminants.find(io)->second[1] * f_inv(L),illuminants.find(io)->second[2] * f_inv(L - b) };
-    }
+    cv::Vec3d tolab(cv::Vec3d& lab);
 
     /* *\ brief Calculate To.
        *\ param src the input array, type of cv::Mat.
        *\ return the output array, type of cv::Mat
     */
-    cv::Mat tosrc(cv::Mat& src)
-    {
-        return channelWise(src, [this](cv::Vec3d a)->cv::Vec3d {return tolab(a); });
-    }
+    cv::Mat tosrc(cv::Mat& src);
 };
 
 /* *\ brief Define Lab_D65_2 and Lab_D50_2.
