+++ modules/mcc/include/opencv2/mcc/utils.hpp
@@ -1,6 +1,6 @@
 // This file is part of OpenCV project.
-// It is subject to the license terms in the LICENSE file found in the top-level directory
-// of this distribution and at http://opencv.org/license.html.
+// It is subject to the license terms in the LICENSE file found in the top-level
+// directory of this distribution and at http://opencv.org/license.html.
 
 /*
  * MIT License
@@ -14,8 +14,8 @@
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
  *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
@@ -29,28 +29,50 @@
 #ifndef __OPENCV_MCC_UTILS_HPP__
 #define __OPENCV_MCC_UTILS_HPP__
 
-#include <functional>
-#include <vector>
-#include <string>
-#include <iostream>
 #include <opencv2/core.hpp>
 
-namespace cv
-{
-namespace ccm
-{
 
-double gammaCorrection_(const double& element, const double& gamma);
-cv::Mat gammaCorrection(const cv::Mat& src, const double& gamma);
-cv::Mat maskCopyTo(const cv::Mat& src, const cv::Mat& mask);
-cv::Mat multiple(const cv::Mat& xyz, const cv::Mat& ccm);
-cv::Mat saturate(cv::Mat& src, const double& low, const double& up);
-cv::Mat rgb2gray(cv::Mat rgb);
+namespace cv {
+namespace ccm {
+
+/* *\ brief gamma correction ,see ColorSpace.pdf for details.
+ *\ param src the input array,type of cv::Mat.
+ *\ param gamma a constant for gamma correction.
+ */
+CV_EXPORTS double gammaCorrection_(const double& element, const double& gamma);
+
+CV_EXPORTS cv::Mat gammaCorrection(const cv::Mat& src, const double& gamma);
+
+/* *\ brief maskCopyTo a function to delete unsatisfied elementwise.
+ *\ param src the input array, type of cv::Mat.
+ *\ param mask operation mask that used to choose satisfided elementwise.
+ */
+CV_EXPORTS cv::Mat maskCopyTo(const cv::Mat& src, const cv::Mat& mask);
+
+/* *\ brief multiple the function used to compute an array with n channels
+ *mulipied by ccm. \ param src the input array, type of cv::Mat. \ param ccm the
+ *ccm matrix to make color correction.
+ */
+CV_EXPORTS cv::Mat multiple(const cv::Mat& xyz, const cv::Mat& ccm);
+
+/* *\ brief multiple the function used to get the mask of saturated colors,
+            colors between low and up will be choosed.
+   *\ param src the input array, type of cv::Mat.
+   *\ param low  the threshold to choose saturated colors
+   *\ param up  the threshold to choose saturated colors
+*/
+CV_EXPORTS cv::Mat saturate(cv::Mat& src, const double& low, const double& up);
+
+/* *\ brief rgb2gray it is an approximation grayscale function for relative RGB
+ *color space, see Miscellaneous.pdf for details; \ param rgb the input array,
+ *type of cv::Mat.
+ */
+CV_EXPORTS cv::Mat rgb2gray(cv::Mat rgb);
 
 /* *\ brief function for elementWise operation
-   *\ param src the input array, type of cv::Mat
-   *\ lambda a for operation
-*/
+ *\ param src the input array, type of cv::Mat
+ *\ lambda a for operation
+ */
 template<typename F>
 cv::Mat elementWise(const cv::Mat& src, F&& lambda)
 {
@@ -103,10 +125,10 @@ cv::Mat channelWise(const cv::Mat& src, F&& lambda)
 }
 
 /* *\ brief function for distance operation.
-   *\ param src the input array, type of cv::Mat.
-   *\ param ref another input array, type of cv::Mat.
-   *\ param lambda the computing method for distance .
-*/
+ *\ param src the input array, type of cv::Mat.
+ *\ param ref another input array, type of cv::Mat.
+ *\ param lambda the computing method for distance .
+ */
 template<typename F>
 cv::Mat distanceWise(cv::Mat& src, cv::Mat& ref, F&& lambda)
 {
@@ -121,117 +143,11 @@ cv::Mat distanceWise(cv::Mat& src, cv::Mat& ref, F&& lambda)
     return dst;
 }
 
-
-double gammaCorrection_(const double& element, const double& gamma)
-{
-    return (element >= 0 ? pow(element, gamma) : -pow((-element), gamma));
-}
-
-/* *\ brief gamma correction ,see ColorSpace.pdf for details.
-   *\ param src the input array,type of cv::Mat.
-   *\ param gamma a constant for gamma correction.
-*/
-cv::Mat gammaCorrection(const cv::Mat& src, const double& gamma)
-{
-    return elementWise(src, [gamma](double element)->double {return gammaCorrection_(element, gamma); });
-}
-
-/* *\ brief maskCopyTo a function to delete unsatisfied elementwise.
-   *\ param src the input array, type of cv::Mat.
-   *\ param mask operation mask that used to choose satisfided elementwise.
- */
-cv::Mat maskCopyTo(const cv::Mat& src, const cv::Mat& mask)
-{
-    cv::Mat dst(countNonZero(mask), 1, src.type());
-    const int channel = src.channels();
-    auto it_mask = mask.begin<uchar>();
-    switch (channel)
-    {
-    case 1:
-    {
-        auto it_src = src.begin<double>(), end_src = src.end<double>();
-        auto it_dst = dst.begin<double>();
-        for (; it_src != end_src; ++it_src, ++it_mask)
-        {
-            if (*it_mask)
-            {
-                (*it_dst) = (*it_src);
-                ++it_dst;
-            }
-        }
-        break;
-    }
-    case 3:
-    {
-        auto it_src = src.begin<cv::Vec3d>(), end_src = src.end<cv::Vec3d>();
-        auto it_dst = dst.begin<cv::Vec3d>();
-        for (; it_src != end_src; ++it_src, ++it_mask)
-        {
-            if (*it_mask)
-            {
-                (*it_dst) = (*it_src);
-                ++it_dst;
-            }
-        }
-        break;
-    }
-    default:
-        throw std::invalid_argument { "Wrong channel!" };
-        break;
-    }
-    return dst;
-}
-
-/* *\ brief multiple the function used to compute an array with n channels mulipied by ccm.
-   *\ param src the input array, type of cv::Mat.
-   *\ param ccm the ccm matrix to make color correction.
-*/
-cv::Mat multiple(const cv::Mat& xyz, const cv::Mat& ccm)
-{
-    cv::Mat tmp = xyz.reshape(1, xyz.rows * xyz.cols);
-    cv::Mat res = tmp * ccm;
-    res = res.reshape(res.cols, xyz.rows);
-    return res;
-}
-
-/* *\ brief multiple the function used to get the mask of saturated colors,
-            colors between low and up will be choosed.
-   *\ param src the input array, type of cv::Mat.
-   *\ param low  the threshold to choose saturated colors
-   *\ param up  the threshold to choose saturated colors
-*/
-cv::Mat saturate(cv::Mat& src, const double& low, const double& up)
-{
-    cv::Mat dst = cv::Mat::ones(src.size(), CV_8UC1);
-    cv::MatIterator_<cv::Vec3d> it_src = src.begin<cv::Vec3d>(), end_src = src.end<cv::Vec3d>();
-    cv::MatIterator_<uchar> it_dst = dst.begin<uchar>();
-    for (; it_src != end_src; ++it_src, ++it_dst)
-    {
-        for (int i = 0; i < 3; ++i)
-        {
-            if ((*it_src)[i] > up || (*it_src)[i] < low)
-            {
-                *it_dst = 0.;
-                break;
-            }
-        }
-    }
-    return dst;
-}
+CV_EXPORTS cv::Mat multiple(const cv::Mat& xyz, const cv::Mat& ccm);
 
 const static cv::Mat m_gray = (cv::Mat_<double>(3, 1) << 0.2126, 0.7152, 0.0722);
 
-/* *\ brief rgb2gray it is an approximation grayscale function for relative RGB color space,
-   *         see Miscellaneous.pdf for details;
-   *\ param rgb the input array, type of cv::Mat.
- */
-cv::Mat rgb2gray(cv::Mat rgb)
-{
-    return multiple(rgb, m_gray);
-}
-
-} // namespace ccm
-} // namespace cv
-
+}  // namespace ccm
+}  // namespace cv
 
 #endif
